requirements.txt
paho-mqtt==1.6.1
python-decouple==3.8
tensorflow==2.13.0rc0
joblib==1.2.0
dill==0.3.6
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
ipykernel==6.23.1
ipython==8.13.2